pyproject.toml
[project]
name = "vision-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "vision-agents[gemini,getstream]>=0.2.2",
    "vision-agents-plugins-ultralytics>=0.2.2",
]
